refactor: placeId를 url에서 받도록 변경 및 place 단일 수정으로 변경
