+++ backend/src/main/java/middle_point_search/backend/domains/place/controller/PlaceController.java
@@ -3,6 +3,7 @@ package middle_point_search.backend.domains.place.controller;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,6 +17,7 @@ import middle_point_search.backend.common.dto.BaseResponse;
 import middle_point_search.backend.common.dto.DataResponse;
 import middle_point_search.backend.domains.place.dto.PlaceDTO;
 import middle_point_search.backend.domains.place.dto.PlaceDTO.PlaceSaveRequest;
+import middle_point_search.backend.domains.place.dto.PlaceDTO.PlaceUpdateRequest;
 import middle_point_search.backend.domains.place.dto.PlaceDTO.PlacesFindResponse;
 import middle_point_search.backend.domains.place.dto.PlaceDTO.PlacesSaveBySelfRequest;
 import middle_point_search.backend.domains.place.service.PlaceService;
@@ -43,10 +45,10 @@ public class PlaceController {
 		return ResponseEntity.ok(BaseResponse.ok());
 	}
 
-	@PutMapping
-	public ResponseEntity<BaseResponse> placeUpdate(@RequestBody PlaceDTO.PlacesUpdateRequest request) {
+	@PutMapping("/{placeId}")
+	public ResponseEntity<BaseResponse> placeUpdate(@PathVariable("placeId") Long placeId, @RequestBody PlaceUpdateRequest request) {
 
-		placeService.updatePlaces(request);
+		placeService.updatePlaces(placeId, request);
 
 		return ResponseEntity.ok(BaseResponse.ok());
 	}
