+++ backend/src/main/java/middle_point_search/backend/domains/place/dto/PlaceDTO.java
@@ -30,28 +30,17 @@ public class PlaceDTO {
 		List<PlaceSaveRequest> addresses;
 	}
 
-
-
 	@Getter
 	@NoArgsConstructor(access = AccessLevel.PRIVATE)
-	public static class PlacesUpdateRequest {
+	public static class PlaceUpdateRequest {
 
-		List<PlaceUpdateRequest> places;
-
-		@Getter
-		@NoArgsConstructor(access = AccessLevel.PRIVATE)
-		public static class PlaceUpdateRequest {
-
-			private Long placeId;
-			private String siDo;
-			private String siGunGu;
-			private String roadNameAddress;
-			private Double addressLat;
-			private Double addressLong;
-			private Transport transport;
-		}
+		private String siDo;
+		private String siGunGu;
+		private String roadNameAddress;
+		private Double addressLat;
+		private Double addressLong;
+		private Transport transport;
 	}
-
 
 	@Getter
 	@AllArgsConstructor(access = AccessLevel.PRIVATE)
